requirements.txt
astroid==3.2.2
asttokens==2.4.1
colorama==0.4.6
coverage==7.5.3
dill==0.3.8
exceptiongroup==1.2.1
executing==2.0.1
Faker==25.8.0
icecream==2.1.3
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
numpy==2.0.0
packaging==24.0
pandas==2.2.2
platformdirs==4.2.2
pluggy==1.5.0
Pygments==2.18.0
pylint==3.2.2
pytest==8.2.1
pytest-cov==5.0.0
pytest-pylint==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.12.0
tzdata==2024.1
